chore: add temporary suite-common/intl-types package
