+++ packages/suite/package.json
@@ -36,6 +36,7 @@
         "@trezor/urls": "*",
         "@trezor/utils": "*",
         "@trezor/validation": "*",
+        "@suite-common/intl-types": "*",
         "bignumber.js": "^9.0.2",
         "date-fns": "^2.28.0",
         "dropbox": "^10.30.0",
